pyproject.toml
[build-system]
requires = ["setuptools", "setuptools_scm"]
build-backend = "setuptools.build_meta"

[project]
name = "mseg"
maintainers = [
  { name = "Lukas Turcani", email = "lukasturcani93@gmail.com" },
]

dependencies = [
    "clasp-segmentation",
    "matplotlib>=3.7.5",
    "numpy<1.24",
    "plotly>=6.0.0",
    "polars>=1.8.2",
    "ruptures>=1.1.9",
    "sklearn-contrib-py-earth",
    "stumpy>=1.13.0",
]
requires-python = ">=3.8,<3.9"
dynamic = ["version"]
readme = "README.rst"
description = "Does things for you."

[dependency-groups]
dev = [
  "ruff",
  "mypy",
  "pytest",
  "pytest-cov",
  "sphinx",
  "sphinx-copybutton",
  "furo",
  "notebook>=7.3.2",
]

[project.urls]
github = "https://github.com/lukasturcani/mseg"
documentation = "https://mseg.readthedocs.io"

[tool.setuptools_scm]

[tool.ruff]
line-length = 79

[tool.ruff.lint]
select = ["ALL"]
ignore = ["ANN101", "ANN401", "COM812", "ISC001"]

[tool.ruff.lint.pydocstyle]
convention = "google"

[tool.ruff.lint.per-file-ignores]
"tests/*" = [
  "D100",
  "D101",
  "D102",
  "D103",
  "D104",
  "D105",
  "D106",
  "D107",
  "S101",
  "INP001",
]
"examples/*" = [
  "INP001",
]
"docs/source/conf.py" = ["D100", "INP001"]

[tool.mypy]
show_error_codes = true
implicit_optional = false
warn_no_return = true
strict_optional = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true
warn_unreachable = true
disallow_any_generics = true

[[tool.mypy.overrides]]
module = [
  "plotly.*",
  "matplotlib.*",
  "pyearth.*",
]
ignore_missing_imports = true

[tool.pytest.ini_options]
addopts = "--cov=src --cov-report term-missing"

[tool.uv.sources]
sklearn-contrib-py-earth = { git = "https://github.com/scikit-learn-contrib/py-earth", rev = "v0.2dev" }
clasp-segmentation = { git = "https://github.com/patrickzib/clasp_segmentation" }
